// schematics/src/store/schema.d.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Creates a new store.
 */
export interface PwaStoreOptionsSchema {
  path?: string;
  restricted?: boolean;
  tsext?: string;
  project?: string;
  /**
   * The name of the store.
   */
  name?: string;
  /**
   * The extension to generate the store in. If empty it is generated in the project core store.
   */
  extension?: string;
  /**
   * The feature to generate the store in the core store. If empty it is generated in the project core store.
   */
  feature?: string;
  /**
   * The entity to generate a ngrx entity state from.
   */
  entity?: string;
}